Add cri subcommand and add ctr/command package
